# openapi/kit-merged.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Realtime Service",
    "version": "1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/realtime/notification/list": {
      "post": {
        "operationId": "NotificationService_ListNotification2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListNotificationReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListNotificationRequest"
            }
          }
        ],
        "tags": [
          "NotificationService"
        ]
      }
    },
    "/v1/realtime/notification/{id}": {
      "get": {
        "operationId": "NotificationService_GetNotification",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Notification"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "NotificationService"
        ]
      },
      "delete": {
        "operationId": "NotificationService_DeleteNotification",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteNotificationReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "NotificationService"
        ]
      }
    },
    "/v1/realtime/notification/{id}/read": {
      "put": {
        "summary": "ReadNotification set notification as read",
        "operationId": "NotificationService_ReadNotification",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "id. set \"-\" for read all",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "tags": [
          "NotificationService"
        ]
      }
    },
    "/v1/realtime/notifications": {
      "get": {
        "operationId": "NotificationService_ListNotification",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListNotificationReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "afterPageToken",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "beforePageToken",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "filter.hasRead.$eq",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "filter.hasRead.$neq",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "filter.hasRead.$null",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "filter.hasRead.$nnull",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "NotificationService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "operationBooleanFilterOperators": {
      "type": "object",
      "properties": {
        "$eq": {
          "type": "boolean",
          "x-nullable": true
        },
        "$neq": {
          "type": "boolean",
          "x-nullable": true
        },
        "$null": {
          "type": "boolean",
          "x-nullable": true
        },
        "$nnull": {
          "type": "boolean",
          "x-nullable": true
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "v1DeleteNotificationReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
...
    "v1ListNotificationReply": {
      "type": "object",
      "properties": {
        "totalSize": {
          "type": "integer",
          "format": "int32"
        },
        "filterSize": {
          "type": "integer",
          "format": "int32"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Notification"
          }
        },
        "nextAfterPageToken": {
          "type": "string",
          "x-nullable": true
        },
        "nextBeforePageToken": {
          "type": "string",
          "x-nullable": true
        },
        "unreadSize": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ListNotificationRequest": {
      "type": "object",
      "properties": {
        "afterPageToken": {
          "type": "string"
        },
        "beforePageToken": {
          "type": "string"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "fields": {
          "type": "string"
        },
        "filter": {
          "$ref": "#/definitions/v1NotificationFilter"
        }
      }
    },
...
    "v1Notification": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "tenantId": {
          "type": "string"
        },
        "group": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "link": {
          "type": "string"
        },
        "source": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "extra": {
          "type": "object"
        },
        "level": {
          "type": "integer",
          "format": "int32"
        },
        "hasRead": {
          "type": "boolean"
        }
      }
    },
    "v1NotificationFilter": {
      "type": "object",
      "properties": {
        "hasRead": {
          "$ref": "#/definitions/operationBooleanFilterOperators"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "description": "Authentication token, prefixed by Bearer: Bearer \u003ctoken\u003e",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}